reorganisation of directories
