+++ mobile-browser-based-version/src/components/building_frames/containers/TrainingFrame.vue
@@ -88,7 +88,7 @@
 </template>
 
 <script>
-import UploadingFrame from "../UploadingFrame";
+import UploadingFrame from "../upload/UploadingFrame";
 import TrainingInformationFrame from "../TrainingInformationFrame";
 import ActionFrame from "./ActionFrame";
 import IconCard from "../../containers/IconCard";
